Bunch of gridworld stuff

+++ gridworld/src/consts/energy.rs
@@ -1,2 +1,2 @@
-pub const CREATURE_START_ENERGY: u16 = 1000;
-pub const MOVE_COST: u16 = 1;
+pub const CREATURE_START_ENERGY: u32 = 1000;
+pub const MOVE_COST: u32 = 1;
